Trackpad changes

+++ logic.h
@@ -37,8 +37,8 @@ typedef struct gamepad_status {
     uint8_t center;
     uint8_t touchpad_press;
 
-    uint8_t touchpadX;
-    uint8_t touchpadY;
+    uint16_t touchpadX;
+    uint16_t touchpadY;
 
     uint8_t l4;
     uint8_t r4;
